components/camera_capture/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    make freeze
#
apscheduler==3.7.0
azure-common==1.1.27
azure-core==1.17.0
azure-storage-blob==12.8.1
boto3==1.18.30
botocore==1.21.30
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
google-api-core==2.0.0
google-auth==2.0.1
google-cloud-core==2.0.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
idna==3.2
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.21
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
smart-open[all]==5.2.0
tzlocal==2.1
urllib3==1.26.6

# The following packages are considered to be unsafe in a requirements file:
# setuptools
